% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_network_data}
\alias{extract_network_data}
\title{Extract data frame from an nma_data object}
\usage{
extract_network_data(network)
}
\arguments{
\item{network}{A network object of class \code{nma_data} (from \code{multinma}) or a list.}
}
\value{
A data.frame containing the network data with standardized names.
}
\description{
Helper function to extract a tidy data frame from a multinma network object.
It handles mapping internal \code{.study}, \code{.trt}, \code{.n}, etc. variables to clean names
and consolidates IPD and/or AgD data.
}
