% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preflight_covariates.R
\name{plot_study_timeline}
\alias{plot_study_timeline}
\title{Plot Study Timeline}
\usage{
plot_study_timeline(
  data,
  study_var = "study",
  trt_var = "treatment",
  tx_duration_var = "val_tx_duration",
  followup_var = "timepoint_wk",
  facet_var = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{study_var}{Column name for study IDs.}

\item{trt_var}{Column name for treatments.}

\item{tx_duration_var}{Column name for treatment duration.}

\item{followup_var}{Column name for total follow-up time.}

\item{facet_var}{Optional column name to facet the timeline plot by.}
}
\value{
A ggplot object.
}
\description{
Visualizes the duration of treatment and follow-up across different studies and arms.
}
