% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma_preflight.R
\name{print.nma_preflight}
\alias{print.nma_preflight}
\title{Print NMA Preflight Summary}
\usage{
\method{print}{nma_preflight}(x, ...)
}
\arguments{
\item{x}{An \code{nma_preflight} object.}

\item{...}{Additional arguments.}
}
\description{
Print NMA Preflight Summary
}
