% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preflight_sparsity.R
\name{print.preflight_sparsity}
\alias{print.preflight_sparsity}
\title{Print Preflight Sparsity Summary}
\usage{
\method{print}{preflight_sparsity}(x, ...)
}
\arguments{
\item{x}{A \code{preflight_sparsity} object.}

\item{...}{Additional arguments.}
}
\description{
Print Preflight Sparsity Summary
}
